refactor: restructure pet components by domain features
