+++ frontend/src/components/common/index.ts
@@ -1,4 +1,5 @@
 export { default as AuthActionButton } from './auth-action-button';
 export { default as TransactionStatus } from './transaction-status';
+export { default as DashboardPanel } from './dashboard-panel';
 // ui primitives re-exported for convenience
 export { NeonButton, NeonCard, NeonModal } from '@components/ui';
