+++ frontend/src/components/pet/creation/pet-creator/index.css
