+++ frontend/src/router/routes/rename-route/index.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import InteractionStandalone from '@components/pet/interaction-standalone';
+import InteractionStandalone from '@components/pet/interactions/standalone';
 import RenamePanel from '@components/pet/interactions/rename-panel';
 
 const RenameRoute: React.FC = () => (
